feat: update name and add more advanced types
