+++ packages/falso/src/index.ts
@@ -201,4 +201,4 @@ export {
   incrementalDate,
   IncrementalDateOptions,
 } from './lib/factories/incremental-date';
-export { randPercentages } from './lib/percentages';
+export { randAggregation } from './lib/aggregation';
